✨ feat: veiculo_sem_an controller/model/route, cracha model, new modulo pessoas (dependente+visitante) 🔧 fix: removed foto from getAllEfetivos for optimization, create veiculo error

+++ backend/src/models/Efetivo.js
@@ -52,7 +52,7 @@ const Efetivo = db.define(
             }
         },
         qrcode_efetivo: {
-            type: DataTypes.INTEGER(7).UNSIGNED.ZEROFILL,
+            type: DataTypes.INTEGER(11).UNSIGNED.ZEROFILL,
             allowNull: true,
             unique: true,
             references: {
